backend/requirements.txt
# Runtime dependencies
fastapi>=0.129.0
uvicorn[standard]>=0.41.0
sqlalchemy>=2.0.0
alembic>=1.18.0
pydantic>=2.12.0
pydantic-settings>=2.13.0
PyJWT[crypto]>=2.8.0
bcrypt>=4.0.0
python-multipart>=0.0.22
